+++ include/chemfiles/formats/XTC.hpp
@@ -26,13 +26,31 @@ class XTCFormat final : public Format {
     size_t nsteps() override;
 
   private:
+    struct FrameHeader {
+        int natoms; /* The total number of atoms */
+        int step;   /* Current step number       */
+        float time; /* Current time              */
+    };
+
+    /// Read header of the Frame at the current position
+    FrameHeader read_frame_header();
+    /// Write header of a Frame
+    void write_frame_header(const FrameHeader& header);
+    /// Determine the number of frames
+    /// and the corresponding offset within the file
+    void determine_frame_offsets();
+
     /// Associated XDR file
     XDRFile file_;
+    /// Offsets within file for fast indexing
+    std::vector<uint64_t> frame_offsets_;
     /// The next step to read
     size_t step_ = 0;
+    /// The number of atoms in the trajectory
+    size_t natoms_ = 0;
 };
 
-template<> const FormatMetadata& format_metadata<XTCFormat>();
+template <> const FormatMetadata& format_metadata<XTCFormat>();
 
 } // namespace chemfiles